revert changes to run_stage
